+++ src/app/(core)/layout.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import NavbarHelper from '@/components/layout/NavbarHelper'
+import PageHeader from '@/components/layout/PageHeader'
 
 interface LayoutProps {
   children: React.ReactNode
@@ -9,11 +10,14 @@ export default function ApplicationLayout({
   children
 }: LayoutProps): JSX.Element {
   return (
-    <main className={'flex bg-gray-100 gap-1'}>
+    <div className={'flex bg-gray-100 w-screen'}>
       {' '}
       {/* 🚨 gap-1 is useful to separate potential overlapping between the primary sidebar and children */}
       <NavbarHelper />
-      {children}
-    </main>
+      <div className={'flex-1 px-4 md:px-6 lg:px-10'}>
+        <PageHeader />
+        {children}
+      </div>
+    </div>
   )
 }
